voteFunction
