--- code/function-app-ri/.gitignore
@@ -1,1 +1,0 @@
-local.settings.json
