--- code/function-app-ri/FunctionApp/host.json
@@ -1,3 +1,0 @@
-{
-    "version": "2.0"
-}
