--- code/function-app-ri/FunctionVoteCounter.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29728.190
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VoteCounter", "FunctionApp\VoteCounter.csproj", "{E7492071-D848-45E3-8153-F8417A967173}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E7492071-D848-45E3-8153-F8417A967173}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E7492071-D848-45E3-8153-F8417A967173}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E7492071-D848-45E3-8153-F8417A967173}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E7492071-D848-45E3-8153-F8417A967173}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {27DFF7C6-FF8F-4B26-B2C6-061C5F59448D}
-	EndGlobalSection
-EndGlobal
