+++ code/function-app-ri/VoteCounter.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.36408.4 d17.14
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VoteCounter", "VoteCounter\VoteCounter.csproj", "{EDD7CFBD-B772-4E4F-95E2-C1824D4BFAB4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EDD7CFBD-B772-4E4F-95E2-C1824D4BFAB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EDD7CFBD-B772-4E4F-95E2-C1824D4BFAB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EDD7CFBD-B772-4E4F-95E2-C1824D4BFAB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EDD7CFBD-B772-4E4F-95E2-C1824D4BFAB4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D2D74D3D-130C-4D90-9F45-766675CCAFC2}
+	EndGlobalSection
+EndGlobal
