+++ code/function-app-ri/VoteCounter/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "VoteCounter": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7021",
+      "launchBrowser": false
+    }
+  }
+}
