+++ code/web-app-ri/VotingData/Properties/launchSettings.json
@@ -1,0 +1,12 @@
+{
+  "profiles": {
+    "VotingData": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "applicationUrl": "https://localhost:55022;http://localhost:55023"
+    }
+  }
+}
